Using linebytes = pixels per byte * width
Instead of total bytes / height. They should technically both be
the same value, but for some reason, they're not. Not sure how
this value is used later for indexing/addressing the framebuffer,
excess bytes in framebuffer go at end of line or at end of buffer?

If at end of buffer, this calculation method is correct. If at end
of line, then it's not. Usually, the two are one and the same, but
definitely not always.