Fixed: Skewed graphics on GPUs that over-allocate the framebuffer
After testing on an AMD Radeon HD7770, for a framebuffer of 800x600x16 in
theory, only 960000 bytes would be needed, rounded up to 963560 for alignment.
However, fb_size for this resolution is actually 998400 because of over-
allocation on the GPU's end. This extra address space is spread out at the end
of each scanline, and must be accounted for.